@@ -4,65 +4,63 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2.5",
+        "php": "^8.0.0",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "abraham/twitteroauth": "^3.0",
-        "api-platform/api-pack": "^1.3",
-        "composer/package-versions-deprecated": "1.11.99.1",
+        "api-platform/core": "^3.2",
         "doctrine/annotations": "^1.0",
+        "doctrine/dbal": "^3",
         "doctrine/doctrine-bundle": "^2.2",
-        "doctrine/doctrine-migrations-bundle": "^3.0",
-        "doctrine/orm": "^2.8",
+        "doctrine/doctrine-migrations-bundle": "^3.2",
+        "doctrine/orm": "^2.14",
         "ifsnop/mysqldump-php": "^2.9",
         "knplabs/knp-paginator-bundle": "^5.3",
-        "lexik/jwt-authentication-bundle": "2.10",
+        "lcobucci/jwt": "*",
+        "lexik/jwt-authentication-bundle": "*",
         "matthiasmullie/minify": "^1.3",
+        "nelmio/cors-bundle": "^2.4",
         "phpdocumentor/reflection-docblock": "^5.2",
-        "sensio/framework-extra-bundle": "^5.1",
-        "seregazhuk/pinterest-bot": "^5.9",
+        "phpstan/phpdoc-parser": "^1.29",
         "spipu/html2pdf": "^5.2",
-        "symfony/asset": "5.2.*",
-        "symfony/console": "5.2.*",
-        "symfony/dotenv": "5.2.*",
-        "symfony/expression-language": "5.2.*",
+        "symfony/asset": "6.3.*",
+        "symfony/console": "6.3.*",
+        "symfony/dotenv": "6.3.*",
+        "symfony/expression-language": "6.3.*",
         "symfony/flex": "^1.3.1",
-        "symfony/form": "5.2.*",
-        "symfony/framework-bundle": "5.2.*",
-        "symfony/http-client": "5.2.*",
-        "symfony/intl": "5.2.*",
-        "symfony/mailer": "5.2.*",
-        "symfony/mime": "5.2.*",
+        "symfony/form": "6.3.*",
+        "symfony/framework-bundle": "6.3.*",
+        "symfony/http-client": "6.3.*",
+        "symfony/intl": "6.3.*",
+        "symfony/mailer": "6.3.*",
+        "symfony/mime": "6.3.*",
         "symfony/monolog-bundle": "^3.1",
-        "symfony/notifier": "5.2.*",
-        "symfony/process": "5.2.*",
-        "symfony/property-access": "5.2.*",
-        "symfony/property-info": "5.2.*",
-        "symfony/proxy-manager-bridge": "5.2.*",
-        "symfony/security-bundle": "5.2.*",
-        "symfony/serializer": "5.2.*",
-        "symfony/string": "5.2.*",
-        "symfony/swiftmailer-bundle": "^3.5",
-        "symfony/translation": "5.2.*",
-        "symfony/twig-bundle": "^5.2",
-        "symfony/uid": "5.2.*",
-        "symfony/validator": "5.2.*",
-        "symfony/web-link": "5.2.*",
+        "symfony/notifier": "6.3.*",
+        "symfony/process": "6.3.*",
+        "symfony/property-access": "6.3.*",
+        "symfony/property-info": "6.3.*",
+        "symfony/proxy-manager-bridge": "6.3.*",
+        "symfony/security-bundle": "6.3.*",
+        "symfony/serializer": "6.3.*",
+        "symfony/string": "6.3.*",
+        "symfony/translation": "6.3.*",
+        "symfony/twig-bundle": "6.3.*",
+        "symfony/uid": "6.3.*",
+        "symfony/validator": "6.3.*",
+        "symfony/web-link": "6.3.*",
         "symfony/webpack-encore-bundle": "^1.11",
-        "symfony/yaml": "5.2.*",
-        "tetranz/select2entity-bundle": "^3.1",
-        "twig/extra-bundle": "^2.12|^3.0",
-        "twig/twig": "^2.12|^3.0"
+        "symfony/yaml": "6.3.*",
+        "tetranz/select2entity-bundle": "*",
+        "twig/extra-bundle": "*"
     },
     "require-dev": {
-        "symfony/browser-kit": "^5.2",
-        "symfony/css-selector": "^5.2",
-        "symfony/debug-bundle": "^5.2",
+        "symfony/browser-kit": "6.3.*",
+        "symfony/css-selector": "6.3.*",
         "symfony/maker-bundle": "^1.0",
-        "symfony/phpunit-bridge": "^5.2",
-        "symfony/stopwatch": "^5.2",
-        "symfony/var-dumper": "^5.2",
-        "symfony/web-profiler-bundle": "^5.2"
+        "symfony/phpunit-bridge": "6.3.*",
+        "symfony/stopwatch": "6.3.*",
+        "symfony/var-dumper": "6.3.*",
+        "symfony/web-profiler-bundle": "6.3.*"
     },
     "config": {
         "optimize-autoloader": true,
