@@ -10,6 +10,7 @@ use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Symfony\Component\Routing\Annotation\Route;
+use Doctrine\ORM\EntityManagerInterface;
 
 /**
  * @Route("/contact")
@@ -29,7 +30,7 @@ class ContactController extends AbstractController
     /**
      * @Route("/send")
      */
-	public function sendAction(Request $request, SessionInterface $session, TranslatorInterface $translator)
+	public function sendAction(EntityManagerInterface $em, Request $request, SessionInterface $session, TranslatorInterface $translator)
 	{
 		$entity = new Contact();
         $form = $this->createForm(ContactType::class, $entity);
@@ -37,9 +38,8 @@ class ContactController extends AbstractController
 
 		if($form->isValid())
 		{
-			$entityManager = $this->getDoctrine()->getManager();
-			$entityManager->persist($entity);
-			$entityManager->flush();
+			$em->persist($entity);
+			$em->flush();
 			$session->getFlashBag()->add('message', $translator->trans("contact.field.YourMessageHasBeenSentSuccessfully"));
 
 			return $this->redirect($this->generateUrl('index'));
