@@ -14,6 +14,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Symfony\Component\Routing\Annotation\Route;
+use Doctrine\ORM\EntityManagerInterface;
 
 /**
  * @Route("/poeticform")
@@ -31,7 +32,7 @@ class PoeticFormAdminController extends AbstractController
     /**
      * @Route("/datatables")
      */
-	public function indexDatatablesAction(Request $request, TranslatorInterface $translator)
+	public function indexDatatablesAction(EntityManagerInterface $em, Request $request, TranslatorInterface $translator)
 	{
 		$iDisplayStart = $request->query->get('iDisplayStart');
 		$iDisplayLength = $request->query->get('iDisplayLength');
@@ -49,9 +50,8 @@ class PoeticFormAdminController extends AbstractController
 			}
 		}
 
-		$entityManager = $this->getDoctrine()->getManager();
-		$entities = $entityManager->getRepository(PoeticForm::class)->getDatatablesForIndex($iDisplayStart, $iDisplayLength, $sortByColumn, $sortDirColumn, $sSearch);
-		$iTotal = $entityManager->getRepository(PoeticForm::class)->getDatatablesForIndex($iDisplayStart, $iDisplayLength, $sortByColumn, $sortDirColumn, $sSearch, true);
+		$entities = $em->getRepository(PoeticForm::class)->getDatatablesForIndex($iDisplayStart, $iDisplayLength, $sortByColumn, $sortDirColumn, $sSearch);
+		$iTotal = $em->getRepository(PoeticForm::class)->getDatatablesForIndex($iDisplayStart, $iDisplayLength, $sortByColumn, $sortDirColumn, $sSearch, true);
 
 		$output = array(
 			"sEcho" => $request->query->get('sEcho'),
@@ -83,11 +83,10 @@ class PoeticFormAdminController extends AbstractController
     /**
      * @Route("/new")
      */
-    public function newAction(Request $request)
+    public function newAction(EntityManagerInterface $em, Request $request)
     {
 		$entity = new PoeticForm();
-		$entityManager = $this->getDoctrine()->getManager();
-		$entity->setLanguage($entityManager->getRepository(Language::class)->findOneBy(["abbreviation" => $request->getLocale()]));
+		$entity->setLanguage($em->getRepository(Language::class)->findOneBy(["abbreviation" => $request->getLocale()]));
         $form = $this->genericCreateForm($request->getLocale(), $entity);
 
 		return $this->render('PoeticForm/new.html.twig', array('form' => $form->createView()));
@@ -96,13 +95,13 @@ class PoeticFormAdminController extends AbstractController
     /**
      * @Route("/login")
      */
-	public function createAction(Request $request, TranslatorInterface $translator)
+	public function createAction(EntityManagerInterface $em, Request $request, TranslatorInterface $translator)
 	{
 		$entity = new PoeticForm();
         $form = $this->genericCreateForm($request->getLocale(), $entity);
 		$form->handleRequest($request);
 		
-		$this->checkForDoubloon($entity, $form);
+		$this->checkForDoubloon($em, $entity, $form);
 
 		if(empty($entity->getFileManagement()) or $entity->getFileManagement()->getPhoto() == null) {
 			$form->get("fileManagement")->get("id")->addError(new FormError($translator->trans("This value should not be blank.", array(), "validators")));
@@ -110,9 +109,8 @@ class PoeticFormAdminController extends AbstractController
 
 		if($form->isValid())
 		{
-			$entityManager = $this->getDoctrine()->getManager();
-			$entityManager->persist($entity);
-			$entityManager->flush();
+			$em->persist($entity);
+			$em->flush();
 
 			$redirect = $this->generateUrl('poeticformadmin_show', array('id' => $entity->getId()));
 
@@ -125,10 +123,9 @@ class PoeticFormAdminController extends AbstractController
     /**
      * @Route("/show/{id}")
      */
-	public function showAction(Request $request, $id)
+	public function showAction(EntityManagerInterface $em, Request $request, $id)
 	{
-		$entityManager = $this->getDoctrine()->getManager();
-		$entity = $entityManager->getRepository(PoeticForm::class)->find($id);
+		$entity = $em->getRepository(PoeticForm::class)->find($id);
 	
 		return $this->render('PoeticForm/show.html.twig', array('entity' => $entity));
 	}
@@ -136,10 +133,9 @@ class PoeticFormAdminController extends AbstractController
     /**
      * @Route("/edit/{id}")
      */
-	public function editAction(Request $request, $id)
+	public function editAction(EntityManagerInterface $em, Request $request, $id)
 	{
-		$entityManager = $this->getDoctrine()->getManager();
-		$entity = $entityManager->getRepository(PoeticForm::class)->find($id);
+		$entity = $em->getRepository(PoeticForm::class)->find($id);
 		$form = $this->genericCreateForm($request->getLocale(), $entity);
 	
 		return $this->render('PoeticForm/edit.html.twig', array('form' => $form->createView(), 'entity' => $entity));
@@ -148,15 +144,14 @@ class PoeticFormAdminController extends AbstractController
     /**
      * @Route("/update/{id}")
      */
-	public function updateAction(Request $request, $id)
+	public function updateAction(EntityManagerInterface $em, Request $request, $id)
 	{
-		$entityManager = $this->getDoctrine()->getManager();
-		$entity = $entityManager->getRepository(PoeticForm::class)->find($id);
+		$entity = $em->getRepository(PoeticForm::class)->find($id);
 
 		$form = $this->genericCreateForm($request->getLocale(), $entity);
 		$form->handleRequest($request);
 		
-		$this->checkForDoubloon($entity, $form);
+		$this->checkForDoubloon($em, $entity, $form);
 
 		if(empty($entity->getFileManagement()) or $entity->getFileManagement()->getPhoto() == null) {
 			$form->get("fileManagement")->get("id")->addError(new FormError($translator->trans("This value should not be blank.", array(), "validators")));
@@ -164,8 +159,8 @@ class PoeticFormAdminController extends AbstractController
 
 		if($form->isValid())
 		{
-			$entityManager->persist($entity);
-			$entityManager->flush();
+			$em->persist($entity);
+			$em->flush();
 
 			$redirect = $this->generateUrl('poeticformadmin_show', array('id' => $entity->getId()));
 
@@ -178,11 +173,10 @@ class PoeticFormAdminController extends AbstractController
     /**
      * @Route("languages")
      */
-	public function languageAction(Request $request)
+	public function languageAction(EntityManagerInterface $em, Request $request)
 	{
-		$entityManager = $this->getDoctrine()->getManager();
 		$locale = $request->query->get("locale");
-		$entities = $entityManager->getRepository(PoeticForm::class)->findAllByLanguage($locale);
+		$entities = $em->getRepository(PoeticForm::class)->findAllByLanguage($locale);
 		
 		$res = array();
 		
@@ -202,12 +196,11 @@ class PoeticFormAdminController extends AbstractController
 		return $this->createForm(PoeticFormType::class, $entity, array('locale' => $locale));
 	}
 
-	private function checkForDoubloon($entity, $form)
+	private function checkForDoubloon(EntityManagerInterface $em, $entity, $form)
 	{
 		if($entity->getTitle() != null)
 		{
-			$entityManager = $this->getDoctrine()->getManager();
-			$checkForDoubloon = $entityManager->getRepository(PoeticForm::class)->checkForDoubloon($entity);
+			$checkForDoubloon = $em->getRepository(PoeticForm::class)->checkForDoubloon($entity);
 
 			if($checkForDoubloon > 0)
 				$form->get("title")->addError(new FormError('Cette entrée existe déjà !'));
