@@ -6,28 +6,31 @@ use Doctrine\ORM\Mapping as ORM;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 
-use ApiPlatform\Core\Annotation\ApiResource;
+use ApiPlatform\Metadata\ApiResource;
+use ApiPlatform\Metadata\Get;
+use ApiPlatform\Metadata\Post;
+use ApiPlatform\Metadata\Put;
+use ApiPlatform\Metadata\Delete;
+use ApiPlatform\Metadata\GetCollection;
 use App\Service\GenericFunction;
 use Symfony\Component\Serializer\Annotation\Groups;
-use ApiPlatform\Core\Annotation\ApiProperty;
-use ApiPlatform\Core\Annotation\ApiSubresource;
+use ApiPlatform\Metadata\ApiProperty;
+use ApiPlatform\Metadata\ApiSubresource;
 
 /**
  * @ORM\Entity(repositoryClass="App\Repository\QuoteRepository")
- * @ApiResource(
- *     itemOperations={
- *          "get"={"security"="is_granted('ROLE_ADMIN')"},
- *          "put"={"security"="is_granted('ROLE_ADMIN')"},
- *          "delete"={"security"="is_granted('ROLE_ADMIN')"}
- *      },
- *     collectionOperations={
- *          "get"={"security"="is_granted('ROLE_ADMIN')"},
- *          "post"={"security"="is_granted('ROLE_ADMIN')"}
- *      },
- *     normalizationContext={"groups"={"read"}},
- *     denormalizationContext={"groups"={"write"}}
- * )
  */
+#[ApiResource(
+    operations: [
+        new Get(security: "is_granted('ROLE_ADMIN')"),
+        new Put(security: "is_granted('ROLE_ADMIN')"),
+		new Delete(security: "is_granted('ROLE_ADMIN')"),
+		new Post(security: "is_granted('ROLE_ADMIN')"),
+        new GetCollection(security: "is_granted('ROLE_ADMIN')")
+    ],
+    normalizationContext: ['groups' => ['read']],
+    denormalizationContext: ['groups' => ['write']]
+)]
 class Quote
 {
 	const BIOGRAPHY_AUTHORTYPE = "biography";
@@ -44,7 +47,7 @@ class Quote
      * @ORM\Id
      * @ORM\GeneratedValue
      * @ORM\Column(type="integer")
-     * @ApiProperty(identifier=false)
+     * #[ApiProperty(identifier: false)]
      */
     protected $id;
 
@@ -61,21 +64,20 @@ class Quote
 
     /**
      * @ORM\Column(type="string", length=500, unique=true)
-     * @ApiProperty(identifier=true)
+     * #[ApiProperty(identifier: true)]
      * @Groups({"read", "write"})
      */
     protected $identifier;
 
 	/**
      * @ORM\ManyToOne(targetEntity="App\Entity\Country")
-     * Groups({"read", "write"})
      */
     protected $country;
 
    /**
     * @ORM\OneToMany(targetEntity=QuoteImage::class, cascade={"persist", "remove"}, mappedBy="quote", orphanRemoval=true)
     * @Groups({"write"})
-	* @ApiSubresource(maxDepth=1)
+	* #[ApiSubresource(maxDepth: 1)]
     */
     protected $images;
 	
